pulls game play functions out of index.js and into world.js
