+++ lib/index.js
@@ -7,12 +7,12 @@ const ctx = canvas.getContext('2d');
 const World = require('./world');
 const HealthFood = require('./health-food');
 const RandomHealthFood = require('./random-health-food');
-const Overlord = require('./overlord');
+// const Overlord = require('./overlord');
 const JunkFood = require('./junk-food');
 const RandomJunkFood = require('./random-junk-food');
-const Gallery = require('./gallery');
-const ScoreBoard = require('./score-board');
-const TrayCollision = require('./tray-collision');
+// const Gallery = require('./gallery');
+// const ScoreBoard = require('./score-board');
+// const TrayCollision = require('./tray-collision');
 
 var gallery;
 var world;
@@ -20,72 +20,75 @@ var overlord;
 var overlordImage;
 var trayCollision;
 // can add new game stuff into world or a renamed world (as game.js)
-function makeNewGame(){
-  gallery = new Gallery();
+// function makeNewGame(){
+  // gallery = new Gallery();
   world = new World(canvas.width, canvas.height, ctx);
-  overlord = new Overlord();
-  overlordImage = new Image();
-  trayCollision = new TrayCollision(world.tray);
-  gallery.start();
-  gameRunning = true;
+  // overlord = new Overlord();
+  // overlordImage = new Image();
+  // trayCollision = new TrayCollision(world.tray);
+  // gallery.start();
+  // gameRunning = true;
   canvas.style.display = 'block';
-  startButton.style.display = 'none';
+  // startButton.style.display = 'none';
   window.setInterval(createHealthFood, 1500);
-}
+  // startScores();
+// }
 
 var healthFoods = [];
 var junkFoods = [];
-var scoreBoard = '';
+// var scoreBoard = '';
 var startButton = document.getElementById('start');
 var reqAnimFrameID = 0;
-var gameRunning = false;
+// var gameRunning = false;
 
-// can move event listener to score-board
-document.addEventListener('DOMContentLoaded', function(){
-  var list = document.getElementById('high-scores');
-  scoreBoard = new ScoreBoard(list);
-  scoreBoard.loadStoredScores();
-});
 
 //game loop- can be stripped but mostly stay in index. Take out what we can to modules.
 startButton.addEventListener('click', function(){
-  audioOn();
-  setGameStyling();
-  makeNewGame();
+  world.audioOn();
+  world.setGameStyling();
+  world.makeNewGame();
   masterGameLoop();
-  userInput();
+  world.userInput();
   increaseDifficulty(overlord, world);
 });
   
-function userInput(){
-  document.addEventListener('keydown', function(event){
-    if (event.keyCode === 39) { world.rightArrow();}
-    if (event.keyCode === 37) { world.leftArrow();}
-  });
-}  
+// function userInput(){
+//   document.addEventListener('keydown', function(event){
+//     if (event.keyCode === 39) { world.rightArrow();}
+//     if (event.keyCode === 37) { world.leftArrow();}
+//   });
+// }  
 function masterGameLoop(){
   requestAnimationFrame(function gameLoop(){
     checkGameEnders();
     beginGame();
-    checkLevel();
+    trayCollision.checkLevel();
   reqAnimFrameID = requestAnimationFrame(gameLoop);
   });
 }
   //pull this into a function
   
-function audioOn(){
-  var gameAudio = document.createElement("audio");
-  gameAudio.src = "Get_Outside.mp3";
-  gameAudio.play();
-  gameAudio.pause();
-}
+// function audioOn(){
+//   var gameAudio = document.createElement("audio");
+//   gameAudio.src = "Get_Outside.mp3";
+//   gameAudio.play();
+//   gameAudio.pause();
+// }
 
-function setGameStyling(){
-  document.getElementById('instructions').style = "display: none";
-  document.getElementById('score-title').style = "display: block";
-  document.getElementById('high-scores').style = "display: block";
-  document.getElementById('level').innerHTML = "<h2>Level 1</h2>";
-}
+// function startScores(){
+//   document.addEventListener('DOMContentLoaded', function(){
+//     var list = document.getElementById('high-scores');
+//     scoreBoard = new ScoreBoard(list);
+//     scoreBoard.loadStoredScores();
+//   });
+// }
+
+// function setGameStyling(){
+//   document.getElementById('instructions').style = "display: none";
+//   document.getElementById('score-title').style = "display: block";
+//   document.getElementById('high-scores').style = "display: block";
+//   document.getElementById('level').innerHTML = "<h2>Level 1</h2>";
+// }
 
 //pull this into a level.js or game.js
 function increaseDifficulty(overlord, world){
@@ -109,14 +112,14 @@ function increaseDifficulty(overlord, world){
   oscillate(overlord, world);
 }
 //pull this into level or game
-function checkLevel(){
-  var level = document.getElementById('level');
-  if(trayCollision.total >= 200 && trayCollision.total < 400 ){
-    level.innerHTML = "<h3>Level 2</h3>";
-  } else if (trayCollision.total >= 400 && trayCollision.total < 600 ){
-    level.innerHTML = "<h3>Level 3</h3>";
-  }
-}
+// function checkLevel(){
+//   var level = document.getElementById('level');
+//   if(trayCollision.total >= 200 && trayCollision.total < 400 ){
+//     level.innerHTML = "<h3>Level 2</h3>";
+//   } else if (trayCollision.total >= 400 && trayCollision.total < 600 ){
+//     level.innerHTML = "<h3>Level 3</h3>";
+//   }
+// }
 function checkGameEnders(){
   if(trayCollision.total < 0 && gameRunning || trayCollision.total >= 600 && gameRunning || !gameRunning){
     endGame();
