+++ lib/tray-collision.js
@@ -12,6 +12,14 @@ TrayCollision.prototype.check = function(foodsArray, instance, reqAnimID){
   }
 };
 
+TrayCollision.prototype.checkLevel = function(){
+  var level = document.getElementById('level');
+  if(this.total >= 200 && this.total < 400 ){
+    level.innerHTML = "<h3>Level 2</h3>";
+  } else if (this.total >= 400 && this.total < 600 ){
+    level.innerHTML = "<h3>Level 3</h3>";
+  }
+};
 
 function isFoodCollidingWithTray(food, tray){
   return !(
